# artifacts/api-server/src/routes/products.ts
import { Router } from "express";
import db, { parseRows, parseOne } from "../lib/db.js";
import { requireAdmin } from "../middlewares/auth.js";
import type { Row } from "../lib/types.js";

const router = Router();

// ─── Public store endpoints ────────────────────────────────────────────────────

router.get("/store/products", (req, res) => {
  const { category, q, limit = "20", page = "1" } = req.query as Record<string, string>;
  let sql = `SELECT * FROM products WHERE status='active'`;
  const params: unknown[] = [];
  if (category) { sql += ` AND category=?`; params.push(category); }
  if (q) { sql += ` AND (title LIKE ? OR tags LIKE ?)`; params.push(`%${q}%`, `%${q}%`); }
  const all = db.prepare(sql).all(...params) as Row[];
  const total = all.length;
  const pageNum = Math.max(1, parseInt(page));
  const limitNum = Math.min(50, Math.max(1, parseInt(limit)));
  const sliced = all.slice((pageNum - 1) * limitNum, pageNum * limitNum);
  const getVariants = db.prepare(`SELECT * FROM variants WHERE product_id=?`);
  const products = parseRows(sliced).map((p) => ({
    ...p,
    variants: parseRows(getVariants.all(p["id"] as string) as Row[]),
  }));
  res.json({ data: products, meta: { total, page: pageNum, limit: limitNum, pages: Math.ceil(total / limitNum) }, error: null });
});

router.get("/store/products/:id", (req, res) => {
  const product = parseOne(db.prepare(`SELECT * FROM products WHERE id=?`).get(req.params["id"]) as Row | undefined);
  if (!product) { res.status(404).json({ data: null, meta: {}, error: "Product not found" }); return; }
  const variants = parseRows(db.prepare(`SELECT * FROM variants WHERE product_id=?`).all(req.params["id"]) as Row[]);
  res.json({ data: { ...product, variants }, meta: {}, error: null });
});

router.get("/store/search", (req, res) => {
  const { q = "" } = req.query as Record<string, string>;
  const rows = db.prepare(`SELECT * FROM products WHERE status='active' AND (title LIKE ? OR vendor LIKE ? OR tags LIKE ?)`).all(`%${q}%`, `%${q}%`, `%${q}%`) as Row[];
  res.json({ data: parseRows(rows), meta: { total: rows.length, query: q }, error: null });
});

router.get("/store/collections", (_req, res) => {
  const rows = db.prepare(`SELECT * FROM collections`).all() as Row[];
  res.json({ data: parseRows(rows), meta: { total: rows.length }, error: null });
});

router.get("/store/collections/:id", (req, res) => {
  const col = parseOne(db.prepare(`SELECT * FROM collections WHERE id=?`).get(req.params["id"]) as Row | undefined);
  if (!col) { res.status(404).json({ data: null, meta: {}, error: "Collection not found" }); return; }
  res.json({ data: col, meta: {}, error: null });
});

router.get("/store/settings", (_req, res) => {
  const row = db.prepare(`SELECT value FROM settings WHERE key='store'`).get() as Row | undefined;
  const s = (row ? JSON.parse(row["value"] as string) : {}) as Record<string, unknown>;
  res.json({
    data: {
      shippingMethods: s["shippingMethods"] ?? [],
      tax: s["tax"] ?? { enabled: false, inclusive: false, regions: [] },
      paymentMethods: s["paymentMethods"] ?? { card: true, cod: true, applePay: false, paypal: false },
      currency: s["currency"] ?? "USD",
    },
    meta: {},
    error: null,
  });
});

// ─── Admin: products ──────────────────────────────────────────────────────────

router.use("/admin/products", requireAdmin);

router.get("/admin/products", (req, res) => {
  const { status, category, q } = req.query as Record<string, string>;
  let sql = `SELECT p.*, (SELECT COALESCE(SUM(v.inventory),0) FROM variants v WHERE v.product_id=p.id) AS total_inventory, (SELECT COUNT(*) FROM variants v WHERE v.product_id=p.id) AS variants_count FROM products p WHERE 1=1`;
  const params: unknown[] = [];
  if (status) { sql += ` AND p.status=?`; params.push(status); }
  if (category) { sql += ` AND p.category=?`; params.push(category); }
  if (q) { sql += ` AND p.title LIKE ?`; params.push(`%${q}%`); }
  const rows = db.prepare(sql).all(...params) as Row[];
  res.json({ data: parseRows(rows), meta: { total: rows.length }, error: null });
});

router.get("/admin/products/:id", (req, res) => {
  const product = parseOne(db.prepare(`SELECT * FROM products WHERE id=?`).get(req.params["id"]) as Row | undefined);
  if (!product) { res.status(404).json({ data: null, meta: {}, error: "Product not found" }); return; }
  const variants = parseRows(db.prepare(`SELECT * FROM variants WHERE product_id=?`).all(req.params["id"]) as Row[]);
  res.json({ data: { ...product, variants }, meta: {}, error: null });
});

router.post("/admin/products", (req, res) => {
  const id = `p_${Date.now()}`;
  const now = new Date().toISOString();
  const b = req.body as Record<string, unknown>;
  db.prepare(`INSERT INTO products (id,title,vendor,category,description,price,compare_price,images,tags,status,created_at,updated_at) VALUES (?,?,?,?,?,?,?,?,?,?,?,?)`)
    .run(id, b["title"] ?? "", b["vendor"] ?? "", b["category"] ?? "women", b["description"] ?? "", b["price"] ?? 0, b["compareAtPrice"] ?? null, JSON.stringify(b["images"] ?? []), JSON.stringify(b["tags"] ?? []), b["status"] ?? "draft", now, now);
  const product = parseOne(db.prepare(`SELECT * FROM products WHERE id=?`).get(id) as Row | undefined);
  res.status(201).json({ data: product, meta: {}, error: null });
});

router.put("/admin/products/:id", (req, res) => {
  const id = req.params["id"];
  const existing = db.prepare(`SELECT id FROM products WHERE id=?`).get(id);
  if (!existing) { res.status(404).json({ data: null, meta: {}, error: "Product not found" }); return; }
  const b = req.body as Record<string, unknown>;
  const now = new Date().toISOString();
  db.prepare(`UPDATE products SET title=COALESCE(?,title), vendor=COALESCE(?,vendor), category=COALESCE(?,category), description=COALESCE(?,description), price=COALESCE(?,price), compare_price=?, images=COALESCE(?,images), tags=COALESCE(?,tags), status=COALESCE(?,status), updated_at=? WHERE id=?`)
    .run(b["title"] ?? null, b["vendor"] ?? null, b["category"] ?? null, b["description"] ?? null, b["price"] ?? null, b["compareAtPrice"] ?? null, b["images"] !== undefined ? JSON.stringify(b["images"]) : null, b["tags"] !== undefined ? JSON.stringify(b["tags"]) : null, b["status"] ?? null, now, id);
  const product = parseOne(db.prepare(`SELECT * FROM products WHERE id=?`).get(id) as Row | undefined);
  res.json({ data: product, meta: {}, error: null });
});

router.delete("/admin/products/:id", (req, res) => {
  const id = req.params["id"];
  if (!db.prepare(`SELECT id FROM products WHERE id=?`).get(id)) { res.status(404).json({ data: null, meta: {}, error: "Product not found" }); return; }
  db.prepare(`DELETE FROM products WHERE id=?`).run(id);
  res.json({ data: { deleted: true }, meta: {}, error: null });
});

// ─── Admin: collections ───────────────────────────────────────────────────────

router.use("/admin/collections", requireAdmin);

router.get("/admin/collections", (_req, res) => {
  const rows = db.prepare(`SELECT * FROM collections`).all() as Row[];
  res.json({ data: parseRows(rows), meta: { total: rows.length }, error: null });
});

router.get("/admin/collections/:id", (req, res) => {
  const col = parseOne(db.prepare(`SELECT * FROM collections WHERE id=?`).get(req.params["id"]) as Row | undefined);
  if (!col) { res.status(404).json({ data: null, meta: {}, error: "Collection not found" }); return; }
  res.json({ data: col, meta: {}, error: null });
});

router.post("/admin/collections", (req, res) => {
  const id = `col_${Date.now()}`;
  const b = req.body as Record<string, unknown>;
  db.prepare(`INSERT INTO collections (id,title,description,image,products_count,created_at) VALUES (?,?,?,?,?,?)`)
    .run(id, b["title"] ?? "", b["description"] ?? "", b["image"] ?? "", 0, new Date().toISOString());
  const col = parseOne(db.prepare(`SELECT * FROM collections WHERE id=?`).get(id) as Row | undefined);
  res.status(201).json({ data: col, meta: {}, error: null });
});

router.put("/admin/collections/:id", (req, res) => {
  const id = req.params["id"];
  if (!db.prepare(`SELECT id FROM collections WHERE id=?`).get(id)) { res.status(404).json({ data: null, meta: {}, error: "Collection not found" }); return; }
  const b = req.body as Record<string, unknown>;
  db.prepare(`UPDATE collections SET title=COALESCE(?,title), description=COALESCE(?,description), image=COALESCE(?,image) WHERE id=?`).run(b["title"] ?? null, b["description"] ?? null, b["image"] ?? null, id);
  const col = parseOne(db.prepare(`SELECT * FROM collections WHERE id=?`).get(id) as Row | undefined);
  res.json({ data: col, meta: {}, error: null });
});

router.delete("/admin/collections/:id", (req, res) => {
  db.prepare(`DELETE FROM collections WHERE id=?`).run(req.params["id"]);
  res.json({ data: { deleted: true }, meta: {}, error: null });
});

// ─── Admin: variants / inventory ──────────────────────────────────────────────

router.use("/admin/variants", requireAdmin);

router.get("/admin/variants", (_req, res) => {
  const rows = db.prepare(`SELECT * FROM variants`).all() as Row[];
  res.json({ data: parseRows(rows), meta: { total: rows.length }, error: null });
});

router.get("/admin/variants/:productId", (req, res) => {
  const rows = db.prepare(`SELECT * FROM variants WHERE product_id=?`).all(req.params["productId"]) as Row[];
  res.json({ data: parseRows(rows), meta: { total: rows.length }, error: null });
});

router.post("/admin/variants", (req, res) => {
  const b = req.body as Record<string, unknown>;
  const productId = b["productId"] as string | undefined;
  if (!productId || !db.prepare(`SELECT id FROM products WHERE id=?`).get(productId)) { res.status(404).json({ data: null, meta: {}, error: "Product not found" }); return; }
  const id = `var_${Date.now()}_${Math.floor(Math.random() * 1000)}`;
  const option1 = (b["option1"] as string | undefined) ?? null;
  const option2 = (b["option2"] as string | undefined) ?? null;
  const derived = [option1, option2].filter(Boolean).join(" / ");
  const title = (b["title"] as string | undefined) ?? (derived || "Default Title");
  db.prepare(`INSERT INTO variants (id,product_id,title,sku,price,compare_price,inventory,option1,option2) VALUES (?,?,?,?,?,?,?,?,?)`)
    .run(id, productId, title, b["sku"] ?? "", b["price"] ?? 0, b["comparePrice"] ?? null, b["inventory"] ?? 0, option1, option2);
  const variant = parseOne(db.prepare(`SELECT * FROM variants WHERE id=?`).get(id) as Row | undefined);
  res.status(201).json({ data: variant, meta: {}, error: null });
});

router.put("/admin/variants/:id/update", (req, res) => {
  const id = req.params["id"];
  if (!db.prepare(`SELECT id FROM variants WHERE id=?`).get(id)) { res.status(404).json({ data: null, meta: {}, error: "Variant not found" }); return; }
  const b = req.body as Record<string, unknown>;
  db.prepare(`UPDATE variants SET inventory=COALESCE(?,inventory), price=COALESCE(?,price), sku=COALESCE(?,sku), title=COALESCE(?,title), compare_price=COALESCE(?,compare_price) WHERE id=?`).run(b["inventory"] ?? null, b["price"] ?? null, b["sku"] ?? null, b["title"] ?? null, b["comparePrice"] ?? null, id);
  const variant = parseOne(db.prepare(`SELECT * FROM variants WHERE id=?`).get(id) as Row | undefined);
  res.json({ data: variant, meta: {}, error: null });
});

router.delete("/admin/variants/:id", (req, res) => {
  const id = req.params["id"];
  if (!db.prepare(`SELECT id FROM variants WHERE id=?`).get(id)) { res.status(404).json({ data: null, meta: {}, error: "Variant not found" }); return; }
  db.prepare(`DELETE FROM variants WHERE id=?`).run(id);
  res.json({ data: { deleted: true }, meta: {}, error: null });
});

export default router;