# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mora e-commerce admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [products]
      security: [{ bearerAuth: [] }]
      summary: List all products
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: adminCreateProduct
      tags: [products]
      security: [{ bearerAuth: [] }]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
  /admin/products/{id}:
    get:
      operationId: adminGetProduct
      tags: [products]
      security: [{ bearerAuth: [] }]
      summary: Get a product with variants
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetailResponse"
    put:
      operationId: adminUpdateProduct
      tags: [products]
      security: [{ bearerAuth: [] }]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
    delete:
      operationId: adminDeleteProduct
      tags: [products]
      security: [{ bearerAuth: [] }]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"

  # ─── Collections ─────────────────────────────────────────────────────────────
  /admin/collections:
    get:
      operationId: adminListCollections
      tags: [collections]
      security: [{ bearerAuth: [] }]
      summary: List collections
      responses:
        "200":
          description: Collection list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionListResponse"
    post:
      operationId: adminCreateCollection
      tags: [collections]
      security: [{ bearerAuth: [] }]
      summary: Create a collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionInput"
      responses:
        "201":
          description: Created collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionResponse"
  /admin/collections/{id}:
    get:
      operationId: adminGetCollection
      tags: [collections]
      security: [{ bearerAuth: [] }]
      summary: Get a collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionResponse"
    put:
      operationId: adminUpdateCollection
      tags: [collections]
      security: [{ bearerAuth: [] }]
      summary: Update a collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionUpdate"
      responses:
        "200":
          description: Updated collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionResponse"
    delete:
      operationId: adminDeleteCollection
      tags: [collections]
      security: [{ bearerAuth: [] }]
      summary: Delete a collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"

  # ─── Variants ────────────────────────────────────────────────────────────────
  /admin/variants:
    get:
      operationId: adminListVariants
      tags: [variants]
      security: [{ bearerAuth: [] }]
      summary: List all variants
      responses:
        "200":
          description: Variant list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VariantListResponse"
    post:
      operationId: adminCreateVariant
      tags: [variants]
      security: [{ bearerAuth: [] }]
      summary: Create a variant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VariantInput"
      responses:
        "201":
          description: Created variant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VariantResponse"
  /admin/variants/{productId}:
    get:
      operationId: adminGetProductVariants
      tags: [variants]
      security: [{ bearerAuth: [] }]
      summary: Get variants for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Variant list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VariantListResponse"
  /admin/variants/{id}/update:
    put:
      operationId: adminUpdateVariant
      tags: [variants]
      security: [{ bearerAuth: [] }]
      summary: Update a variant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VariantUpdate"
      responses:
        "200":
          description: Updated variant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VariantResponse"
  /admin/variants/{id}:
    delete:
      operationId: adminDeleteVariant
      tags: [variants]
      security: [{ bearerAuth: [] }]
      summary: Delete a variant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"

  # ─── Orders ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    DeletedResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            deleted:
              type: boolean
          required: [deleted]
        meta:
          type: object
        error:
          type: ["string", "null"]
      required: [data, meta, error]
    ListMeta:
      type: object
      properties:
        total:
          type: integer
      required: [total]

    # ─── Product ──────────────────────────────────────────────────────────────
    Product:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        vendor: { type: string }
        category: { type: string }
        description: { type: string }
        price: { type: number }
        comparePrice: { type: ["number", "null"] }
        images: { type: array, items: { type: string } }
        tags: { type: array, items: { type: string } }
        status: { type: string }
        totalInventory: { type: ["integer", "null"] }
        variantsCount: { type: ["integer", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, title, vendor, category, price, status]
    Variant:
      type: object
      properties:
        id: { type: string }
        productId: { type: string }
        title: { type: string }
        sku: { type: string }
        price: { type: number }
        comparePrice: { type: ["number", "null"] }
        inventory: { type: integer }
        option1: { type: ["string", "null"] }
        option2: { type: ["string", "null"] }
      required: [id, productId, title, sku, price, inventory]
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            variants:
              type: array
              items:
                $ref: "#/components/schemas/Variant"
          required: [variants]
    ProductInput:
      type: object
      properties:
        title: { type: string }
        vendor: { type: string }
        category: { type: string }
        description: { type: string }
        price: { type: number }
        compareAtPrice: { type: ["number", "null"] }
        images: { type: array, items: { type: string } }
        tags: { type: array, items: { type: string } }
        status: { type: string }
      required: [title, vendor, category, price]
    ProductUpdate:
      type: object
      properties:
        title: { type: string }
        vendor: { type: string }
        category: { type: string }
        description: { type: string }
        price: { type: number }
        compareAtPrice: { type: ["number", "null"] }
        images: { type: array, items: { type: string } }
        tags: { type: array, items: { type: string } }
        status: { type: string }
    VariantInput:
      type: object
      properties:
        productId: { type: string }
        title: { type: string }
        sku: { type: string }
        price: { type: number }
        comparePrice: { type: ["number", "null"] }
        inventory: { type: integer }
        option1: { type: ["string", "null"] }
        option2: { type: ["string", "null"] }
      required: [productId]
    VariantUpdate:
      type: object
      properties:
        inventory: { type: integer }
        price: { type: number }
        sku: { type: string }
        title: { type: string }
        comparePrice: { type: ["number", "null"] }
    ProductListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        meta:
          $ref: "#/components/schemas/ListMeta"
        error:
          type: ["string", "null"]
      required: [data, meta, error]
    ProductResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Product"
        meta:
          type: object
        error:
          type: ["string", "null"]
      required: [data, meta, error]
    ProductDetailResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/ProductDetail"
        meta:
          type: object
        error:
          type: ["string", "null"]
      required: [data, meta, error]
    VariantListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Variant"
        meta:
          $ref: "#/components/schemas/ListMeta"
        error:
          type: ["string", "null"]
      required: [data, meta, error]
    VariantResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Variant"
        meta:
          type: object
        error:
          type: ["string", "null"]
      required: [data, meta, error]

    # ─── Collection ───────────────────────────────────────────────────────────
    Collection:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        image: { type: string }
        productsCount: { type: integer }
        createdAt: { type: string }
      required: [id, title]
    CollectionInput:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        image: { type: string }
      required: [title]
    CollectionUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        image: { type: string }
    CollectionListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Collection"
        meta:
          $ref: "#/components/schemas/ListMeta"
        error:
          type: ["string", "null"]
      required: [data, meta, error]
    CollectionResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Collection"
        meta:
          type: object
        error:
          type: ["string", "null"]
      required: [data, meta, error]

    # ─── Order ────────────────────────────────────────────────────────────────
...