# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mora e-commerce admin API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  title?: string;
  vendor?: string;
  category?: string;
  description?: string;
  price?: number;
  /** @nullable */
  compareAtPrice?: number | null;
  images?: string[];
  tags?: string[];
  status?: string;
}