# lib/api-zod/src/generated/types/variantInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mora e-commerce admin API
 * OpenAPI spec version: 0.1.0
 */

export interface VariantInput {
  productId: string;
  title?: string;
  sku?: string;
  price?: number;
  /** @nullable */
  comparePrice?: number | null;
  inventory?: number;
  /** @nullable */
  option1?: string | null;
  /** @nullable */
  option2?: string | null;
}